Move DTLS 1.2 tests into a dedicated test file
